+++ crates/ide-assists/src/handlers/convert_tuple_struct_to_named_struct.rs
@@ -104,7 +104,7 @@ fn edit_struct_def(
                 ast::make::tokens::single_newline().text(),
             );
             edit.insert(tuple_fields_text_range.start(), w.syntax().text());
-            if !w.syntax().last_token().is_some_and(|t| t.kind() == SyntaxKind::COMMA) {
+            if w.syntax().last_token().is_none_or(|t| t.kind() != SyntaxKind::COMMA) {
                 edit.insert(tuple_fields_text_range.start(), ",");
             }
             edit.insert(
